requirements.txt
# Core Backend
flask==3.0.2
flask-cors==4.0.0

# Machine Learning & Scientific Computing          # For RL model
numpy==1.26.4          # Required by multiple components
scipy==1.12.0          # For entropy calculations
pandas==2.2.1          # For data processing in sim_utils.py
tqdm==4.66.0           # For progress bars in RL model

# AI/LLM Integration
google-generativeai==0.3.2  # For Gemini integration

# Environment & Configuration
python-dotenv==1.0.1   # For loading environment variables

# Utilities
typing-extensions==4.10.0  # For type hints
pathlib==1.0.1         # For file path handling
